+++ .editorconfig
@@ -13,10 +13,12 @@ indent_style = space
 indent_size = 4
 ktlint_function_naming_ignore_when_annotated_with = Composable
 ktlint_standard_blank-line-before-declaration = disabled
+ktlint_class_signature_rule_force_multiline_when_parameter_count_greater_or_equal_than = 1
+ktlint_standard_class-signature = enabled
 max_line_length = 100
 
 [**/test/**/*Test.kt]
 max_line_length = 140
 
-[**/generated/**]
-ktlint_standard = disabled
+[*.toml]
+max_line_length = 140
