+++ .github/workflows/puntsenlinea-ci.yml
@@ -1,0 +1,38 @@
+name: Punts Enlinia CI
+
+on:
+  pull_request:
+    branches: [ main ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: set up JDK 17
+        uses: actions/setup-java@v4
+        with:
+          java-version: '17'
+          distribution: 'zulu'
+          cache: gradle
+
+      - name: Setup Android SDK
+        uses: android-actions/setup-android@v3
+
+      - name: Grant execute permission for gradlew
+        run: chmod +x gradlew
+
+      - name: Ktlint check
+        run: ./gradlew ktlintCheck
+
+      - name: Clean
+        run: ./gradlew clean
+
+      - name: Build with Gradle
+        run: ./gradlew build
+
+      - name: Unit tests
+        run: ./gradlew testDebugUnitTest
