+++ app/src/main/kotlin/com/maasbodev/puntsenlinia/MainActivity.kt
@@ -14,34 +14,37 @@ import androidx.compose.ui.tooling.preview.Preview
 import com.maasbodev.puntsenlinia.ui.theme.PuntsEnliniaTheme
 
 class MainActivity : ComponentActivity() {
-	override fun onCreate(savedInstanceState: Bundle?) {
-		super.onCreate(savedInstanceState)
-		enableEdgeToEdge()
-		setContent {
-			PuntsEnliniaTheme {
-				Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-					Greeting(
-						name = "Android",
-						modifier = Modifier.padding(innerPadding)
-					)
-				}
-			}
-		}
-	}
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            PuntsEnliniaTheme {
+                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                    Greeting(
+                        name = "Android",
+                        modifier = Modifier.padding(innerPadding),
+                    )
+                }
+            }
+        }
+    }
 }
 
 @Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-	Text(
-		text = "Hello $name!",
-		modifier = modifier
-	)
+fun Greeting(
+    name: String,
+    modifier: Modifier = Modifier,
+) {
+    Text(
+        text = "Hello $name!",
+        modifier = modifier,
+    )
 }
 
 @Preview(showBackground = true)
 @Composable
 fun GreetingPreview() {
-	PuntsEnliniaTheme {
-		Greeting("Android")
-	}
+    PuntsEnliniaTheme {
+        Greeting("Android")
+    }
 }
