+++ gradle/libs.versions.toml
@@ -38,3 +38,5 @@ google-services = { id = "com.google.gms.google-services", version = "4.4.2" }
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
 kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
 
+# Quality Plugins
+ktlint = { id = "org.jlleitschuh.gradle.ktlint", version = "12.1.1" }
