+++ settings.gradle.kts
@@ -1,22 +1,22 @@
 pluginManagement {
-	repositories {
-		google {
-			content {
-				includeGroupByRegex("com\\.android.*")
-				includeGroupByRegex("com\\.google.*")
-				includeGroupByRegex("androidx.*")
-			}
-		}
-		mavenCentral()
-		gradlePluginPortal()
-	}
+    repositories {
+        google {
+            content {
+                includeGroupByRegex("com\\.android.*")
+                includeGroupByRegex("com\\.google.*")
+                includeGroupByRegex("androidx.*")
+            }
+        }
+        mavenCentral()
+        gradlePluginPortal()
+    }
 }
 dependencyResolutionManagement {
-	repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
-	repositories {
-		google()
-		mavenCentral()
-	}
+    repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
+    repositories {
+        google()
+        mavenCentral()
+    }
 }
 
 rootProject.name = "Punts Enlinia"
